refactor: ExpIncRequestDto, ExpIncService

+++ src/main/java/com/bovo/Bovo/modules/rewards/dto/request/ExpIncRequestDto.java
@@ -14,8 +14,8 @@ import lombok.Setter;
 public class ExpIncRequestDto {
 
     @NotNull(message = "user_id는 필수 항목입니다.")
-    private Long userId;
+    private Integer userId;
 
     @NotNull(message = "mission_id는 필수 항목입니다.")
-    private Long missionId;
+    private Integer missionId;
 }
